python/requirements.txt
gpxpy==1.6.2
paho_mqtt==2.1.0
zeroconf==0.132.2
